Crate sc_transaction_graph

Generic Transaction Pool

The pool is based on dependency graph between transactions and their priority. The pool is able to return an iterator that traverses transaction graph in the correct order taking into account priorities and dependencies.

Re-exports

pub use self::base_pool::Transaction;

Modules

base_pool

A basic version of the dependency graph.

watcher

Extrinsics status updates.

Structs

IsValidator

A closure that returns true if the local node is a validator that can author blocks.

Options

Pool configuration options.

Pool

Extrinsics pool that performs validation.

Enums

ValidatedTransaction

Pre-validated transaction. Validated pool only accepts transactions wrapped in this enum.

Traits

ChainApi

Concrete extrinsic validation and query logic.

Type Definitions

BlockHash

Block hash type for a pool.

EventStream

Modification notification event stream type;

ExtrinsicFor

Extrinsic type for a pool.

ExtrinsicHash

Extrinsic hash type for a pool.

NumberFor

Block number type for the ChainApi

TransactionFor

A type of transaction stored in the pool