Struct sc_transaction_graph::base_pool::Transaction

pub struct Transaction<Hash, Extrinsic> {
    pub data: Extrinsic,
    pub bytes: usize,
    pub hash: Hash,
    pub priority: Priority,
    pub valid_till: Longevity,
    pub requires: Vec<Tag>,
    pub provides: Vec<Tag>,
    pub propagate: bool,
    pub source: Source,
}

Immutable transaction

Fields
data: Extrinsic

Raw extrinsic representing that transaction.

bytes: usize

Number of bytes encoding of the transaction requires.

hash: Hash

Transaction hash (unique)

priority: Priority

Transaction priority (higher = better)

valid_till: Longevity

At which block the transaction becomes invalid?

requires: Vec<Tag>

Tags required by the transaction.

provides: Vec<Tag>

Tags that this transaction provides.

propagate: bool

Should that transaction be propagated.

source: Source

Source of that transaction.

Implementations

impl<Hash: Clone, Extrinsic: Clone> Transaction<Hash, Extrinsic>

pub fn duplicate(&self) -> Self

Explicit transaction clone.

Transaction should be cloned only if absolutely necessary && we want every reason to be commented. That’s why we Transaction is not Clone, but there’s explicit duplicate method.